refactor: split onboard runtime shell
